separate functions and main process
