+++ ml/python_script/create_model.py
@@ -10,6 +10,8 @@ import matplotlib.pyplot as plt
 import pathlib
 import csv
 
+from functions import *
+
 
 model_type = sys.argv[1]
 if model_type != "classification" and model_type != "certification":
@@ -17,53 +19,6 @@ if model_type != "classification" and model_type != "certification":
 
 RAILS_ROOT = str(pathlib.Path(__file__).resolve().parents[2])
 
-
-def read_csv(filename):
-    user_id = []
-    image_paths = []
-
-    with open(filename) as f:
-        reader = csv.reader(f)
-
-        for row in reader:
-            user_id.append(row[0])
-            image_paths.append(row[1])
-
-    return user_id, image_paths
-
-def preprocess_image(image):
-    image_4ch = tf.image.decode_image(image, channels=4, expand_animations=False)
-    image_4ch = tf.image.resize(image_4ch, [128, 128])
-    image = tf.cast(tf.reduce_sum(image_4ch, 2, keepdims=True), tf.float32)
-    image /= 255.0
-
-    return image
-
-def load_and_preprocess_image(path):
-    image = tf.io.read_file(path)
-    return preprocess_image(image)
-
-def split(list, slice_ratio):
-    slice_index = int(len(list) * slice_ratio)
-    return list[:slice_index], list[slice_index:]
-
-def separate_image_and_label(list):
-    images = []
-    labels = []
-    for i in range(len(list)):
-        images.append(list[i]["image"])
-        labels.append(list[i]["label"])
-    return images, labels
-
-def separate_2images_and_label(list):
-    images1 = []
-    images2 = []
-    labels = []
-    for i in range(len(list)):
-        images1.append(list[i]["image1"])
-        images2.append(list[i]["image2"])
-        labels.append(list[i]["label"])
-    return images1, images2, labels
 
 def conv_batchnorm_relu(x, filters, kernel_size):
     params = {
@@ -249,7 +204,7 @@ def make_certification_model(images, labels):
     model.fit(
         {"image1": train_images1, "image2": train_images2},
         train_labels,
-        epochs=20,
+        epochs=10,
         validation_data=({"image1": validation_images1, "image2": validation_images2}, validation_labels)
     )
     model.save(RAILS_ROOT + "/ml/hisseki_certification.tf")
@@ -260,4 +215,4 @@ images = list(map(load_and_preprocess_image, image_paths))
 
 model = make_model(images, labels)
 
-print("done")
+print("create_model.py: done")
